@@ -3,19 +3,19 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@testing-library/jest-dom": "^4.2.4",
-    "@testing-library/react": "^9.3.2",
-    "@testing-library/user-event": "^7.1.2",
-    "@types/jest": "^24.0.0",
-    "@types/node": "^12.0.0",
-    "@types/react": "^16.9.0",
-    "@types/react-dom": "^16.9.0",
+    "@testing-library/jest-dom": "5.3.0",
+    "@testing-library/react": "10.0.2",
+    "@testing-library/user-event": "10.0.1",
+    "@types/jest": "25.1.5",
+    "@types/node": "13.11.0",
+    "@types/react": "16.9.32",
+    "@types/react-dom": "16.9.6",
     "classnames": "2.2.6",
-    "react": "^16.13.0",
-    "react-dom": "^16.13.0",
+    "react": "16.13.1",
+    "react-dom": "16.13.1",
     "react-router-dom": "5.1.2",
-    "react-scripts": "3.4.0",
-    "typescript": "~3.7.2"
+    "react-scripts": "3.4.1",
+    "typescript": "3.8.3"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -41,6 +41,6 @@
   "devDependencies": {
     "@types/classnames": "2.2.10",
     "@types/react-router-dom": "5.1.3",
-    "prettier": "1.19.1"
+    "prettier": "2.0.2"
   }
 }
