@@ -3,25 +3,27 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@testing-library/jest-dom": "^4.2.4",
-    "@testing-library/react": "^9.3.2",
-    "@testing-library/user-event": "^7.1.2",
-    "@types/jest": "^24.0.0",
-    "@types/node": "^12.0.0",
-    "@types/react": "^16.9.0",
-    "@types/react-dom": "^16.9.0",
+    "@testing-library/jest-dom": "5.5.0",
+    "@testing-library/react": "10.0.2",
+    "@testing-library/user-event": "10.0.1",
+    "@types/jest": "25.2.1",
+    "@types/node": "13.11.1",
+    "@types/react": "16.9.34",
+    "@types/react-dom": "16.9.6",
     "axios": "0.19.2",
     "classnames": "2.2.6",
-    "react": "^16.13.0",
-    "react-dom": "^16.13.0",
-    "react-scripts": "3.4.0",
-    "typescript": "~3.7.2"
+    "inversify": "5.0.1",
+    "react": "16.13.1",
+    "react-dom": "16.13.1",
+    "react-scripts": "3.4.1",
+    "reflect-metadata": "0.1.13",
+    "typescript": "3.8.3"
   },
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
+    "start": "react-app-rewired start",
+    "build": "react-app-rewired build",
+    "test": "react-app-rewired test",
+    "eject": "react-app-rewired eject"
   },
   "eslintConfig": {
     "extends": "react-app"
@@ -39,7 +41,11 @@
     ]
   },
   "devDependencies": {
+    "@babel/plugin-proposal-decorators": "7.8.3",
     "@types/classnames": "2.2.10",
-    "prettier": "1.19.1"
+    "babel-plugin-parameter-decorator": "1.0.16",
+    "customize-cra": "0.9.1",
+    "prettier": "2.0.4",
+    "react-app-rewired": "2.1.5"
   }
 }
