@@ -4,18 +4,20 @@ export const ExerciseThree: React.FC = () => {
   const [value, setValue] = useState('')
 
   const handleChange = (characters: string) => {
-    if (['a', 'e', 'i', 'o', 'u'].every(vowel => !characters.toLowerCase().includes(vowel))) {
-      setValue(characters)
-    }
+    const charactersWithoutVowels = characters
+      .split('')
+      .filter(character => !['a', 'e', 'i', 'o', 'u'].includes(character))
+      .join('')
+    setValue(charactersWithoutVowels)
   }
 
   return (
-    <div>
-      {value}
-      <label>
-        Input
-        <input type="text" onChange={event => handleChange(event.target.value)} value={value} />
-      </label>
-    </div>
+    <input
+      aria-label="Input"
+      placeholder="Enter characters"
+      type="text"
+      onChange={event => handleChange(event.target.value)}
+      value={value}
+    />
   )
 }
